GetUserInfo更新
（1）学生用户可以查询申请次数、请假次数、工作总时长、累计收入；非学生用户返回0
（2）解决以上值为null时的报错问题

+++ backend/Responses/UserInfoResponse.cs
@@ -18,9 +18,9 @@ namespace SyaApi.Responses
         public int nof_absent { get; set; }
 
         /// 工作总时长
-        public int work_time { get; set; }
+        public double work_time { get; set; }
 
         /// 已获得薪水总量
-        public int income { get; set; }
+        public double income { get; set; }
     }
 }
